Rework asking the user to reroll
Basically, wrapping the read_values() return with an option so I can
handle when nothing is entered in, and therefore progressing the round to
scoring when nothing is entered in. Less clumsy than asking if the user
wants to roll. I've also tried cleaning up my binary so not as much is in
main(). Still more to do, but it's a start.